Accommodating autoincrementing primary keys
requires slightly different schemas for publisher and subscriber.

The publisher can autoincrement, but the subscriber cannot. The
schema for the subscriber primary key needs to be the same type
though.

It may be possible to use the same schema, with an additional
SQL command to set an autoincrement on an existing column. In
fact, GPT says I can do this, so I'll see if I can get that
suggestion running and merge the schema back. I can also
us the test schema which is already in the replication script.