Sequence creation now working and replicating
Following on from yesterday's progress, adding an
autoincrement sequence after table creation is fine
as long as the schema for publisher and subscriber
match. The sequence is added only to the publisher,
the subscriber will acquire the appropriate values.

This change is only the small demo data.

Still to come: delete and update examples.