Manual replication configuration
Upgraded localhost to postges 16, then the container to 16
as well, the point being to manually configure the settings
for logical replication. Everything succeeded with stock
settings and a single ALTER SYSTEM for logical replication
on the localhost server. Replication proceeded with no other
configuration.

Upcoming is altering settings from the scripts instead of
manually. Restarting everything might be tricky.