Publisher replication filters added to books
One the goals of the project was to learn how to filter
table data to different subscribers, which is achieved in
this change. Leadership and technical topics are published
independently.

Container restarting was also tuned: there is no need to
explicitly remove a container when it is started with the
`--rm` flag.