GEN-197 dummy data for influx testing
Two scripts for adding some dummy data show how to
push data into influxdb, and how to extract it. The
data is generated by a bash script emitted (with some
coaxing) from GPT.

Of interest is we're passing in the timestamp, which
is something to note when collecting metric data from
docker containers in the future.