GEM-201 add more data to influxdb
This is an incremental change for adding influx data
from Ruby. The main result is that I don't really have
any idea how influx works, nor how to get the data in
or out which I am interested in. Will need to do more
reading. The mechanics seem to be simple, but there is
a conceptual element I am missing.