GEN-202 Line Protocol: Writing influx data by the numbers (#9)
Since the previous commits have proven I don't
understand the structure of the data going into
Influx, it was time to read the documentation. Influx
implements what they call a "line protocol" which is
very well documented in their online help.

The method write_demo3 provides a good example of
writing influx data in a manner which can be queried
effectively. The key was reading the influx documentation.
GPT did not provide enough context for me to generate a
useful schema, hence the the previous couple of days
thrashing around.

The influx insertion code needs some major cleanup.