GEN-207 Update sample.rb script to query postgres and write to influx (#14)
https://doolin.atlassian.net/browse/GEN-207

This is a difficult set of changes as neither the postgres
querying nor the influx writing was working. It probably
should have been decomposed into 3 separate commits: the
first to rename, the second to get a viable postgres query,
and the third to make a viable influx write.

Note that the acceptance for these changes does not specify
_coorectly_ querying and writing.

A script for running `pgbench` was added to help generate
some data date for the postgres locks query.

This PR is honestly a real disaster. If this were going to
production, I'd split it into at least 2 and possibly more
commits, probably on different tickets. As it stands, the
sampler can write dummy data into influx, and can acquire
locks data from the publisher database running in docker,
but it isn't hooked up yet to write the acquired data.