GEN-209 Extend PGSampler to query indefinitely (#16)
https://doolin.atlassian.net/browse/GEN-209

The Ruby script for sampling locks was extended to
loop until terminated. This allows much easier visualization
of the benchmarking operations. The default sleep is 100
milliseconds, which will need to be configurable in the
future.

The vision is being able to script a complete suite of
load tests which will write out a report summarizing
the results.