GEN-212 Parameterize docker memory in restart script (#17)
https://doolin.atlassian.net/browse/GEN-212

Memory for docker containers for publisher and
subscribers can be set from the command line of
the restart.sh script. If the value isn't present
the default value is 512m and automatically set.

A bit of clean up on the docker command arguments
was done for readability.

Commented out directives for an ELK stack added
for future reference.