GEN-211 Break the publisher db or container
https://doolin.atlassian.net/browse/GEN-211

The idea is breaking postgres in a docker container
will provide insight into why postgres is breaking in
production.

For this particular change, the focus is on parameterization
of the existing scripts, as it's very difficult to keep
track of the parameter configuration inducing an error.
A large amount of the changes here are commented code
which will be progressively implemented in following
changes. GEN-211 will then be revisited to record a
breaking configuration.