GEN-217 Hook up the Docker container logs to Influx
Container logs get connected via the Telegraph tool
for InfluxDB. The conecpt it simple, with two caveats
when putting it into practice:

1. Most of the information is for running on a linux systems,
which is not the same as running it on Macos. Specificallt,
permissions for the socket need to allow writing.

2. The environment variables for Influx need to be passed in
to the container via Docker run.

It would be nice to find a way to remember both of these
as it took many, many hours to figure it out.