GEN-218 Configure Telegraf to run in docker group (#21)
Proof concept enabled with socket chmod 777, which is
nasty and insecure. These changes rectify ownership
by creating a docker group and adding users to it.
All the details are in the associated script. A revert
script is provised as well.

Manual verification shows container activity in Grafana
just as it should. That is, Telegraf is acquiring container
logs and sending them to InfluxDB. Very nice.