GEN-360 Script shared_preload_libraries for pghero convenience
https://doolin.atlassian.net/browse/GEN-360

Modified from GPT:

In general: shared_preload_libraries is a configuration
parameter in PostgreSQL that specifies a list of shared
libraries to be loaded into global memory when the
PostgreSQL server starts, and available to all databases
running on the server. In this case, we need to enable
pg_stat_statements for examining query oerformance.