GEN-366 Refactor pg_sampler for class length
https://doolin.atlassian.net/browse/GEN-366

This class is overdue for a refactor.
Sadly, rubocop's insistence of a 100 line
count max is just too limiting here. There
is not much opportunity for metaprogramming,
and I prefer the clarity of the exisitng
implementation over attempting to isolate
calls to postgres and influx. The class is
essentially a dispatcher, so it makes sense
to keep all the methods associated with
dispatching toegther.

I could define a class for each type of
query, and I may do that in the future,
but it seems overly abstract for now.

Some reek improvements were added as well.